Long -> Integer

+++ seats/src/main/java/com/example/seat/SeatRepository.java
@@ -2,6 +2,6 @@ package com.example.seat;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface SeatRepository extends PagingAndSortingRepository<Seat, Long> {
+public interface SeatRepository extends PagingAndSortingRepository<Seat, Integer> {
 
 }
